Update REMOVE BLOAT.bat
**THIS SCRIPT BREAKS THINGS, DO SO AT DEVICE RISK**

YOU WILL NEED TO REFLASH STOCK IF YOU WANT TO UNDO THIS.

This is for STOCK firmware.

Added some new entries missed the second time. This does break a few things.

So far what breaks for sure is LenovoID and the Factory Reset button in settings.

Other than that everything left behind still functions as it should.
Highly recommend running "fastboot -w" in fastboot afterwards.